create helloworld module
The hellomodule module will print "my module1 Hello world!" on init
and "mymodule1e Clean up Hello World. \n" on exit, as written in the
new file helloworld.c. Kconfig and Makefile will be associated to
define how it should be handled while compiling, as per the docs.
The Kconfig and Makefile in ./driver will point to the new module,
therefore the module will be under Device Driver Section.

Signed-off-by: Khalid Masum <khalid.masum.92@gmail.com>