fscache: fix misdocumented parameter name
The functions fscache_use_cookie and fscache_unuse_cookie have a
parameter named cookie. But they are documented with the name "object".
Which generates warning when creating docs. This commit will replace
the documentation with a better one named "cookie". This new documentation
line is taken from the function fscache_update_cookie in the same
file.

Signed-off-by: Khalid Masum <khalid.masum.92@gmail.com>